Refactored code & changed file structure
